# Faster`substract_bf16x32_genoa` mixed-precision subtraction