@MarkReedZ 2024-09-12
The permutex was 9.2.   LOL I spent a half hour trying to do it your way and completely borked on the blend missing how simple this whole thing was.

```
    d_f32_even.ivec = _mm512_srli_epi32(d_f32_even.ivec, 16);
    d.ivec = _mm512_mask_blend_epi16(0x55555555, d_f32_odd.ivec, d_f32_even.ivec);

-----------------------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------------------
l2sq_bf16_genoa_128d/min_time:10.000/threads:1       9.20 ns         9.19 ns   1000000000 abs_delta=7.80224m
```